pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labforge"
version = "0.1.0"
description = "Spin up isolated threat research labs locally with Docker"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
labforge = "labforge.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/labforge"]
